+++ AP_Spoof/setup.sh
@@ -3,7 +3,6 @@
 ifconfig
 read -p '[*]Enter network adapter to use: ' netadp
 ifconfig $netadp 10.0.0.1/24 up
-ifconfig $netadp
 echo '[*]Adapter IP: 10.0.0.1'
 read -p "PRESS ENTER TO CONTINUE" e
 sudo apt-get install dnsmasq
@@ -22,13 +21,14 @@ echo 10.0.0.2 www.fakewebsite.com >> fakehosts.conf
 echo [*]Creating hostapd.conf file...
 echo interface=$netadp > hostapd.conf
 echo driver=nl80211 >> hostapd.conf
-echo ssid=KITT-AP >> hostapd.conf
+read -p "[*]SSID Name: " ssid
+echo ssid=$ssid >> hostapd.conf
 echo channel=1 >> hostapd.conf
 echo [*]Configuring IPtables
 sudo sysctl -w net.ipv4.ip_forward=1
 sudo iptables -P FORWARD ACCEPT
 sudo iptables --table nat -A POSTROUTING -o wlan0 -j MASQUERADE
-echo "[*]Open 3 windows, 1 for dnsmasq, 2 for hostapd, and 3 if you want to edit fakehosts.conf")
+echo "[*]Open 3 windows, 1 for dnsmasq, 2 for hostapd, and 3 if you want to edit fakehosts.conf"
 echo "[*]1.)dnsmasq -C dnsmasq.conf -H fakehosts.conf -d"
 echo "[*]2.)hostapd ./hostapd.conf"
 echo "[*]3.)./iptables.sh when done to reverse configurations"